requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pywin32==311; sys_platform == 'win32'
openai>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
slowapi==0.1.9
Jinja2==3.1.4